rere

+++ ImbossibleBoss/IB_Frame/Function.h
@@ -35,7 +35,7 @@ static std::string GetFileName(std::string path) {
 
 static std::string TCHARToString(const TCHAR* ptsz)
 {
-	int len = wcslen((wchar_t*)ptsz);
+	size_t len = wcslen((wchar_t*)ptsz);
 	char* psz = new char[2 * len + 1];
 	wcstombs(psz, (wchar_t*)ptsz, 2 * len + 1);
 	std::string s = psz;
@@ -47,7 +47,7 @@ static TCHAR* StringToTCHAR(std::string& s)
 {
 	std::string tstr;
 	const char* all = s.c_str();
-	int len = 1 + strlen(all);
+	size_t len = 1 + strlen(all);
 	wchar_t* t = new wchar_t[len];
 	if (NULL == t) throw std::bad_alloc();
 	mbstowcs(t, all, len);
